requirements.txt
# CANNECT.AI CV Analytics Dependencies

# Core CV Libraries
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# YOLO & Deep Learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
pandas>=2.0.0
scipy>=1.10.0

# API & Networking
requests>=2.31.0
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0

# Optional: Face analysis for demographics
# deepface>=0.0.79
